Refactor admin routes into dedicated dashboard and audit handlers
