+++ debug-admin-access.html
@@ -88,7 +88,7 @@
             log('🔍 Testing API access...', 'info')
             
             try {
-                const response = await fetch('/api/admin?action=dashboard', {
+                const response = await fetch('/api/admin/dashboard', {
                     headers: {
                         'Authorization': `Bearer ${currentSession.access_token}`,
                         'Content-Type': 'application/json'
